缺陷[#84489]: 【声纹设备台账】操作删除时增加判断，已绑定到测点生成巡视点的声纹设备，无法删除
[描述说明]: 删除时增加判断.
[问题原因]:
[解决方法]:
[测试建议]:

+++ iot-center-platform/src/main/java/com/yjh/platform/module/device/service/TVoiceDeviceService.java
@@ -86,6 +86,10 @@ public class TVoiceDeviceService{
 
     @Transactional(rollbackFor = Exception.class)
     public int deleteByPrimaryId(Long voiceDeviceId) {
+        List<Long> insList = tVoiceDeviceDao.selectHaveIns(voiceDeviceId);
+        if (!insList.isEmpty()){
+           throw new BusinessException("该声纹设备已绑定巡视点！");
+        }
         VoiceDeviceAllInfoDetail voiceDeviceInfoDetail = this.tVoiceDeviceDao.selectByPrimaryId(voiceDeviceId);
         this.tVoiceDeviceDao.deleteConf(voiceDeviceInfoDetail.getConfigId());
         String recordKey = "is_record_open_state:"+String.valueOf(voiceDeviceId);
